Certify Yourself Improvements
Working on displaying thumbnails of blank certificates instead of drop
downs, Adding hidden email to me when a certificate is generated.

Next steps will be to display full sized certificates, Add pdf download
options.

After that, I want to add a link to Cafepress or Zazzle to print
certificates on t-shirts or coffee mugs…